Add rule for SHA224 hash in PHP, Java, Go, Python, Ruby (#3542)
SHA224 and other hashes with 224 bits are currently on the edge of what is considered a secure hash function. The Australian Cyber Security Center and NIST have indicated that SHA224 is going to be deprecated in the future.

This adds a rule to detect SHA224 and similar hashes. It is distinct from the existing MD5 and SHA1 rules, because those are currently a security problem, and SHA224 is more of a future compliance problem.

Co-authored-by: Claudio <claudio@r2c.dev>